refactor to move out static method in interchange adaptor to common
